WASM: Fix System.IO.FileSystem.DriveInfo
